requirements.txt
APScheduler==3.6.3
certifi==2020.11.8
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
EasyProcess==0.3
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.0
pytz==2020.4
PyVirtualDisplay==1.3.2
selenium==3.141.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.1
